Table Definitions Added
Added table definitions which will auto-create table if they do not exist.
Added an errors log property to the storage class
Added a sql log property to the storage class
Changed the storage->record() method to return an array not a record object. This prevent alot of confusion and the possibility of creating infinate loops.
Added a niot.js placeholder file in assets/js
Updated Bootstrap to include the storage class
Added prop() method to the record class which sets values without committing the changes in the event that auto-commit is on. This is a protected function.

Tested getting and setting records with the test file which were successful
Setting records was tested both with and without auto-commit

Tested auto-creating the table with the table definition file in place which was successful
Both items are in and working